feat: filesystem pin store migration
