+++ migrate/utilities.go
@@ -1,15 +1,9 @@
 package migrate
 
 import (
-	"log"
 	"math"
-	"path"
-	"sync/atomic"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/service/s3"
 	ds "github.com/ipfs/go-datastore"
 	dshelp "github.com/ipfs/go-ipfs-ds-help"
 )
@@ -52,46 +46,4 @@ func blockToCid(key string) (string, error) {
 func isCidPinned(cid string) bool {
 	pinSuccess, found := pinMap.Load(cid)
 	return found && pinSuccess.(bool)
-}
-
-func s3Paginator(bucket, prefix string) *s3.ListObjectsV2Paginator {
-	cfg, err := config.LoadDefaultConfig(ctx)
-	if err != nil {
-		log.Fatal(err)
-	}
-	s3Client := s3.NewFromConfig(cfg)
-	return s3.NewListObjectsV2Paginator(s3Client, &s3.ListObjectsV2Input{
-		Bucket: aws.String(bucket),
-		Prefix: aws.String(prefix),
-	})
-}
-
-func cidsFromS3Page(page *s3.ListObjectsV2Output) []string {
-	cids := make([]string, 0, page.KeyCount)
-	pageKeysFound := uint32(0)
-	pageKeysConverted := uint32(0)
-	pagePinsRemaining := uint32(0)
-	for _, object := range page.Contents {
-		if object.Size != 0 { // filter out directories
-			pageKeysFound++
-			// Convert file name to CID
-			key := aws.ToString(object.Key)
-			_, name := path.Split(key)
-			cid, err := blockToCid(name)
-			if err == nil {
-				pageKeysConverted++
-				if !isCidPinned(cid) {
-					pagePinsRemaining++
-					cids = append(cids, cid)
-				}
-			} else {
-				log.Printf("convert failed: key=%s, err:%s", key, err)
-			}
-		}
-	}
-
-	log.Printf("keys found=%d, total found=%d", pageKeysFound, atomic.AddUint32(&keysFoundCount, pageKeysFound))
-	log.Printf("keys converted=%d, total converted=%d", pageKeysConverted, atomic.AddUint32(&keysConvertedCount, pageKeysConverted))
-	log.Printf("cids not pinned=%d, total remaining=%d", pagePinsRemaining, atomic.AddUint32(&pinsRemainingCount, pagePinsRemaining))
-	return cids
 }
